backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.3
mangum==0.17.0
boto3==1.34.34
